fix unsigned byte to signed byte convert issue in woff output writer

+++ src/main/java/org/mabb/fontverter/woff/WoffOutputStream.java
@@ -64,7 +64,8 @@ public class WoffOutputStream extends FontDataOutputStream {
             binary = StringUtils.repeat("0", 6 - binary.length()) + binary;
 
         binary = transBinary + binary;
-        byte byteOn = (Byte.parseByte(binary, 2));
+        byte byteOn = (byte) (Integer.parseInt(binary, 2));
+
         write(byteOn);
     }
